backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
anthropic>=0.40.0
yfinance>=0.2.40
pandas>=2.0.0
supabase>=2.4.0
stripe>=8.0.0
httpx>=0.27.0
